# backend/microbill-authentication-service/src/main/resources/openapi/auth-api.yaml
openapi: 3.0.3
info:
  title: MicroBill Authentication API
  description: Authentication and authorization service with JWT
  version: 1.0.0
servers:
  - url: http://localhost:8081/api
    description: Auth Service
  - url: http://localhost:8082/api
    description: Management Service
  - url: http://localhost:8083/api
    description: Billing Service
  - url: http://localhost:8084/api
    description: AI Service
...
paths:
...
  /ai/recommendations:
    post:
      tags:
        - AI
      summary: Get product recommendations
      security:
        - BearerAuth: []
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Recommendations retrieved successfully
        '401':
          description: Unauthorized
  /ai/anomaly-score:
    post:
      tags:
        - AI
      summary: Detect anomalies in invoice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnomalyRequest'
      responses:
        '200':
          description: Anomaly detection completed
        '401':
          description: Unauthorized
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    AnomalyRequest:
      type: object
      required:
        - customerId
        - total
        - items
      properties:
        customerId:
          type: integer
          format: int64
        total:
          type: number
          format: double
        items:
          type: array
          items:
            type: object